+++ CHANGES.md
@@ -289,3 +289,4 @@
 ### Pre-release (work-in-progress)
 - Using the same `OS` string from `General` section of `config.ini` for all checkers
 - Quick Scan: minor display adjustment
+- Phonak: fix critical bug on checking files from latest Target version